--- src/facts_experiment_builder/resources/__init__.py
@@ -1,1 +1,0 @@
-"""Bundled resources (clues YAMLs, etc.)."""
